requirements.txt
Flask
Flask-SQLAlchemy
pytest
pytest-mock
python-dotenv
psycopg2